Adiciona suporte a imagens e melhorias gerais
- Atualiza `LoginDto` para inicializar propriedades com valores padrão.
- Adiciona a propriedade `ImageLocal` em `CategoryDto`, `CreateCategoryDto` e `UpdateCategoryDto`.
- Altera o tipo de `Price` de `double` para `decimal` em `ProductDto` e DTOs relacionados.
- Atualiza `CategoryService` para suportar `ImageLocal` em criação e atualização de categorias.
- Adiciona a propriedade `ImageLocal` na entidade `Category`.
- Inclui um novo diretório `/Itens de Solução/` no arquivo `.slnx`.
- Adiciona um arquivo `README.md` com documentação detalhada do projeto.

+++ AtelieDosPontinhos.Application/DTOs/ProductDto.cs
@@ -12,7 +12,7 @@ namespace AtelieDosPontinhos.Application.DTOs
         public string CoverImageUrl { get; set; } = string.Empty;
         public int CategoryId { get; set; }
         public string CategoryName { get; set; } = string.Empty;
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
         public string Type_Product { get; set; } = string.Empty;
     }
@@ -23,7 +23,7 @@ namespace AtelieDosPontinhos.Application.DTOs
         public string Description { get; set; } = string.Empty;
         public string CoverImageUrl { get; set; } = string.Empty;
         public int CategoryId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
     }
     //atulizar produto
@@ -33,7 +33,7 @@ namespace AtelieDosPontinhos.Application.DTOs
         public string Description { get; set; } = string.Empty;
         public string CoverImageUrl { get; set; } = string.Empty;
         public int CategoryId { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public int Stock { get; set; }
     }
 }
